feat(CS-21): IOL Service

See merge request m2892/financial-monitor!12

+++ Fuentes/Servidor/messirveinvertir-be/app/Controllers/Http/AccountsController.ts
@@ -2,41 +2,36 @@ import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Account from 'App/Models/Account'
 
 export default class AccountsController {
-
   public async listAccounts({ auth, request }: HttpContextContract) {
-    const page = request.input('page',1)
-    const limit = request.input('limit',10)
+    const page = request.input('page', 1)
+    const limit = request.input('limit', 10)
     const orderById = request.input('orderById')
-    
+
     return await Account.query()
       .where('uid', auth.user!.uid)
       .orderBy('id', orderById)
       .paginate(page, limit)
   }
 
-   public async createAccounts({ auth, request }: HttpContextContract) {
+  public async createAccounts({ auth, request }: HttpContextContract) {
     const accounts = request.input('accounts')
 
     return await accounts.map(async (account) => {
-        let newAccount = await Account.create({
-          uid: auth.user!.uid,
-          name: account.name,
-          type: account.type,
-          data: JSON.parse('{}')
-        })
+      let newAccount = await Account.create({
+        uid: auth.user!.uid,
+        name: account.name,
+        type: account.type,
+        data: JSON.parse('{}'),
+      })
 
-        return newAccount.id
-      }
-    )
+      return newAccount.id
+    })
   }
 
-   public async deleteAccounts({ auth, request, response }: HttpContextContract) {
+  public async deleteAccounts({ auth, request, response }: HttpContextContract) {
     const accountIds = request.input('accounts')
 
-    await Account.query()
-      .where('uid',auth.user!.uid)
-      .whereIn('id', accountIds)
-      .delete()
+    await Account.query().where('uid', auth.user!.uid).whereIn('id', accountIds).delete()
 
     return response.noContent
   }
